transcription.py
import youtube_dl
import requests
from google.cloud import speech
from ibm_watson import SpeechToTextV1
from ibm_cloud_sdk_core.authenticators import IAMAuthenticator
import azure.cognitiveservices.speech as azure_speech
import boto3
import openai

def transcribe_video(video_path, service='google'):
    if service == 'google':
        return transcribe_with_google(video_path)
    elif service == 'ibm':
        return transcribe_with_ibm(video_path)
    elif service == 'azure':
        return transcribe_with_azure(video_path)
    elif service == 'amazon':
        return transcribe_with_amazon(video_path)
    elif service == 'openai':
        return transcribe_with_openai(video_path)
    else:
        raise ValueError("Invalid service option. Please choose one of the available options.")

def transcribe_with_google(video_path):
    # Set up the Google Cloud Speech-to-Text client
    client = speech.SpeechClient()

    # Configure transcription request
    config = speech.RecognitionConfig(
        encoding=speech.RecognitionConfig.AudioEncoding.LINEAR16,
        sample_rate_hertz=16000,
        language_code="en-US",
    )
    audio = speech.RecognitionAudio(uri=video_path)

    # Perform the transcription using Google Cloud Speech-to-Text
    response = client.recognize(config=config, audio=audio)

    # Extract the transcriptions from the response
    transcriptions = [result.alternatives[0].transcript for result in response.results]

    # Join the transcriptions into a single string
    transcription_text = " ".join(transcriptions)

    return transcription_text

def transcribe_with_ibm(video_path):
    # Configure IBM Watson Speech-to-Text service
    authenticator = IAMAuthenticator('YOUR_IBM_API_KEY')
    service = SpeechToTextV1(authenticator=authenticator)
    service.set_service_url('YOUR_IBM_SERVICE_URL')

    # Perform the transcription using IBM Watson Speech-to-Text
    with open(video_path, 'rb') as audio_file:
        response = service.recognize(
            audio=audio_file,
            content_type='audio/mp3',  # Adjust the content type based on the video format
            model='en-US_NarrowbandModel',
            max_alternatives=1
        ).get_result()

    # Extract the transcriptions from the response
    transcriptions = [result['alternatives'][0]['transcript'] for result in response['results']]

    # Join the transcriptions into a single string
    transcription_text = " ".join(transcriptions)

    return transcription_text

def transcribe_with_azure(video_path):
    # Configure Azure Speech-to-Text service
    speech_config = azure_speech.SpeechConfig(subscription="YOUR_AZURE_SUBSCRIPTION_KEY", region="YOUR_AZURE_REGION")

    # Create an audio configuration
    audio_config = azure_speech.AudioConfig(filename=video_path)

    # Create a speech recognizer
    recognizer = azure_speech.SpeechRecognizer(speech_config=speech_config, audio_config=audio_config)

    # Perform the transcription using Azure Speech-to-Text
    result = recognizer.recognize_once()

    # Extract the transcriptions from the result
    transcription_text = result.text

    return transcription_text

def transcribe_with_amazon(video_path):
    # Configure Amazon Transcribe service
    transcribe = boto3.client('transcribe', region_name='YOUR_AWS_REGION')

    # Start the transcription job
    response = transcribe.start_transcription_job(
        TranscriptionJobName='transcription_job',
        LanguageCode='en-US',
        Media={'MediaFileUri': video_path},
        OutputBucketName='YOUR_S3_BUCKET_NAME'
    )

    # Wait for the transcription job to complete
    transcribe.get_waiter('transcription_job_completed').wait(TranscriptionJobName='transcription_job')

    # Get the transcription results
    response = transcribe.get_transcription_job(TranscriptionJobName='transcription_job')

    # Retrieve the transcript
    transcription_text = response['TranscriptionJob']['Transcript']['TranscriptFileUri']

    return transcription_text

def transcribe_with_openai(video_path):
    # Load the OpenAI Whisper model
    model = openai.WhisperModel()

    # Perform the transcription using OpenAI Whisper
    # ...

    # Return the transcription text
    return transcription_text
